+++ logic/getComments_test.go
@@ -76,7 +76,7 @@ func TestSimple(t *testing.T) {
 
 	firstTree := commentResponse.Comments[0]
 
-	assert.Equal(t, 1, firstTree.Comment.Id, "expected the first root comment to have the id 1")
+	assert.Equal(t, int64(1), firstTree.Comment.Id, "expected the first root comment to have the id 1")
 	assert.Equal(t, 0, len(firstTree.Children), "expected the first root comment to not have any children")
 }
 
@@ -113,8 +113,8 @@ func TestTwoRootComments(t *testing.T) {
 	firstTree := commentResponse.Comments[0]
 	secondTree := commentResponse.Comments[1]
 
-	assert.Equal(t, 1, firstTree.Comment.Id, "expected the first comment to have the id 1")
-	assert.Equal(t, 2, secondTree.Comment.Id, "expected the second comment to have the id 2")
+	assert.Equal(t, int64(1), firstTree.Comment.Id, "expected the first comment to have the id 1")
+	assert.Equal(t, int64(2), secondTree.Comment.Id, "expected the second comment to have the id 2")
 	assert.Equal(t, 0, len(firstTree.Children), "expected the first root comment to not have any children")
 	assert.Equal(t, 0, len(secondTree.Children), "expected the second root comment to not have any children")
 }
@@ -151,12 +151,12 @@ func TestTwoChainedComments(t *testing.T) {
 	require.Equal(t, 1, len(commentResponse.Comments), "expected the tree to have 1 root comment")
 	firstTree := commentResponse.Comments[0]
 
-	assert.Equal(t, 1, firstTree.Comment.Id, "expected the first root comment to have the id 1")
+	assert.Equal(t, int64(1), firstTree.Comment.Id, "expected the first root comment to have the id 1")
 
 	require.Equal(t, 1, len(firstTree.Children), "expected the first root comment to have 1 child")
 	firstChild := firstTree.Children[0]
 
-	assert.Equal(t, 2, firstChild.Comment.Id, "expected first child to have the id 2")
+	assert.Equal(t, int64(2), firstChild.Comment.Id, "expected first child to have the id 2")
 	assert.Equal(t, 0, len(firstChild.Children), "expected the first child to not have any children")
 }
 
@@ -211,17 +211,17 @@ func TestTwoChains(t *testing.T) {
 	firstTree := commentResponse.Comments[0]
 	secondTree := commentResponse.Comments[1]
 
-	assert.Equal(t, 1, firstTree.Comment.Id, "expected the first root comment to have the id 1")
-	assert.Equal(t, 3, secondTree.Comment.Id, "expected the second root comment to have the id 3")
+	assert.Equal(t, int64(1), firstTree.Comment.Id, "expected the first root comment to have the id 1")
+	assert.Equal(t, int64(3), secondTree.Comment.Id, "expected the second root comment to have the id 3")
 
 	require.Equal(t, 1, len(firstTree.Children), "expected the first root comment to have 1 child")
 	require.Equal(t, 1, len(secondTree.Children), "expected the second root comment to have 1 child")
 	firstChild := firstTree.Children[0]
 	secondChild := secondTree.Children[0]
 
-	assert.Equal(t, 2, firstChild.Comment.Id, "expected first child to have the id 2")
+	assert.Equal(t, int64(2), firstChild.Comment.Id, "expected first child to have the id 2")
 	assert.Equal(t, 0, len(firstChild.Children), "expected the first child to not have any children")
-	assert.Equal(t, 4, secondChild.Comment.Id, "expected second child to have the id 4")
+	assert.Equal(t, int64(4), secondChild.Comment.Id, "expected second child to have the id 4")
 	assert.Equal(t, 0, len(secondChild.Children), "expected the second child to not have any children")
 }
