+++ model/Comment.go
@@ -1,11 +1,11 @@
 package model
 
 type Comment struct {
-	Id          int    `json:"comment_id"`
+	Id          int64  `json:"comment_id"`
 	Author      string `json:"author" binding:"required"`
 	Email       string `json:"email" binding:"required"`
 	Text        string `json:"text" binding:"required"`
-	ParentId    int    `json:"parent_id" binding:""`
+	ParentId    int64  `json:"parent_id" binding:""`
 	CreatedAt   int64  `json:"created_at"`
 	TouchedAt   int64  `json:"touched_at"`
 	NumChildren int    `json:"num_children"`
@@ -16,7 +16,7 @@ type CommentCreation struct {
 	Author     string `json:"author" binding:"required"`
 	Email      string `json:"email" binding:"required"`
 	Text       string `json:"text" binding:"required"`
-	ParentId   int    `json:"parent_id"` // optional, maps to 0 if not provided
+	ParentId   int64  `json:"parent_id"` // optional, maps to 0 if not provided
 }
 
 type CommentTree struct {
@@ -33,7 +33,7 @@ type CommentsResponse struct {
 }
 
 type Thread struct {
-	Id       int
+	Id       int64
 	Path     string
 	NumTotal int
 	NumRoot  int
